requirements.txt
PyOpengl==3.1.6
numpy==1.22.1
esper==2.1
glfw==2.5.5
matplotlib==3.5.0
pillow==9.2.0
mazelib==0.9.13
PyGLM==2.5.7
pygame==2.1.2
git+https://github.com/petrikvladimir/pyphysx.git@master
pywavefront
moderngl-window
moderngl